backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.11.0
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
click==8.1.8
coverage==7.13.0
cryptography==44.0.2
dj-database-url==3.1.0
Django==5.2.4
django-allauth==65.0.2
django-anymail==12.0
django-cors-headers==4.5.0
django-coverage-plugin==3.1.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-filter==25.2
django-model-utils==5.0.0
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
django-redis==5.4.0
django-rest-framework-nested==0.0.1
django-stubs==5.1.0
django-stubs-ext==5.1.1
djangorestframework==3.15.2
djangorestframework-gis==1.2.0
djangorestframework-stubs==3.15.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
drf-yasg==1.21.11
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
inflection==0.5.1
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
phonenumbers==9.0.21
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
python-dotenv==1.2.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
rest-framework-simplejwt==0.0.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.5
starlette==0.46.2
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.34.0
whitenoise==6.11.0
